# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Membership Management API Documentation",
        "description": "API documentation using Swagger",
        "version": "1.0"
    },
    "paths": {
...
        "/api/industries": {
            "get": {
                "tags": [
                    "Industries"
                ],
                "summary": "Lấy danh sách ngành nghề",
                "operationId": "c0a4ae08f3bf3eddc64c4eedca45b955",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên ngành nghề",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số lượng ngành nghề trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách ngành nghề",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Industry"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Industries"
                ],
                "summary": "Thêm ngành nghề",
                "operationId": "dba8b526f449b79e0ffab59d040be74b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreIndustryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ngành nghề đã được tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Industry"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/industries/{id}": {
            "get": {
                "tags": [
                    "Industries"
                ],
                "summary": "Lấy thông tin ngành nghề",
                "operationId": "43f07eaa50d2049f332fe26ba7ff5f31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của ngành nghề",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin ngành nghề",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Industry"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ngành nghề không tồn tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Industries"
                ],
                "summary": "Cập nhật ngành nghề",
                "operationId": "808a4e476b3e49110a6cc60879521d4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của ngành nghề",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreIndustryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật ngành nghề thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Industry"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ngành nghề không tồn tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Industries"
                ],
                "summary": "Xóa ngành nghề",
                "operationId": "5488eb2ba64ff2e4ae9b9855ee5ceb28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của ngành nghề",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa ngành nghề thành công"
                    },
                    "404": {
                        "description": "Ngành nghề không tồn tại hoặc không thể xóa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreIndustryRequest": {
                "title": "Store Industry Request",
                "description": "Yêu cầu tạo mới ngành nghề",
                "required": [
                    "industry_code",
                    "industry_name"
                ],
                "properties": {
                    "industry_code": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "IT"
                    },
                    "industry_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Công nghệ thông tin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ngành liên quan đến phần mềm, phần cứng và công nghệ",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Industry": {
                "title": "Industry",
                "description": "Thông tin ngành nghề",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "industry_code": {
                        "type": "string",
                        "example": "IT"
                    },
                    "industry_name": {
                        "type": "string",
                        "example": "Công nghệ thông tin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ngành liên quan đến phần mềm, phần cứng và công nghệ"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}